# docs/api-spec/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend service for eev",
        "title": "eev-backend",
        "contact": {
            "name": "Alen Alex",
            "url": "http://alenalex.me",
            "email": "contact@alenalex.me"
        },
        "version": "1.0"
    },
    "host": "e.alenalex/me",
    "basePath": "/api/",
    "paths": {
...
        "/share": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Creates a new shareable resource (text, url, or file). Requires authentication via access_token cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shareable"
                ],
                "summary": "Create a new shareable",
                "parameters": [
                    {
                        "description": "Create shareable request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateShareableRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Shareable created",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateShareableResponse"
                        }
                    },
                    "202": {
                        "description": "Shareable created - awaiting encrypted data",
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateShareableResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/share/{code}": {
            "get": {
                "description": "Retrieves public shareable information using its unique code.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "shareable"
                ],
                "summary": "Get shareable by code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shareable code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shareable info",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetShareableResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid code",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Shareable not found",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.CreateShareableRequest": {
            "type": "object",
            "required": [
                "name",
                "time_params",
                "type"
            ],
            "properties": {
                "allowed_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "data": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ShareableFileRequest"
                    }
                },
                "name": {
                    "type": "string"
                },
                "notification_params": {
                    "$ref": "#/definitions/handlers.CreateShareableRequestNotificationParams"
                },
                "options": {
                    "$ref": "#/definitions/handlers.CreateShareableRequestParams"
                },
                "time_params": {
                    "$ref": "#/definitions/handlers.CreateShareableRequestTimeParams"
                },
                "type": {
                    "$ref": "#/definitions/services.ShareableType"
                }
            }
        },
        "handlers.CreateShareableRequestNotificationParams": {
            "type": "object",
            "properties": {
                "email_notification_on_open": {
                    "type": "boolean"
                },
                "notify_target_emails": {
                    "type": "boolean"
                }
            }
        },
        "handlers.CreateShareableRequestParams": {
            "type": "object",
            "properties": {
                "encrypt": {
                    "type": "boolean"
                },
                "only_once": {
                    "type": "boolean"
                }
            }
        },
        "handlers.CreateShareableRequestTimeParams": {
            "type": "object",
            "properties": {
                "active_from": {
                    "type": "string"
                },
                "expiry_at": {
                    "type": "string"
                }
            }
        },
        "handlers.CreateShareableResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "uploads": {
                    "description": "only present for file type",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ShareableFileUpload"
                    }
                }
            }
        },
        "handlers.GetShareableResponse": {
            "type": "object",
            "properties": {
                "active_from": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "expiry_at": {
                    "type": "string"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.ShareableFileResponse"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "options": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "type": {
                    "type": "string"
                },
                "user_email": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ShareableFileRequest": {
            "type": "object",
            "required": [
                "content_type",
                "file_name",
                "file_size"
            ],
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                }
            }
        },
        "handlers.ShareableFileResponse": {
            "type": "object",
            "properties": {
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "share_id": {
                    "type": "string"
                },
                "signed_url": {
                    "type": "string"
                }
            }
        },
        "handlers.ShareableFileUpload": {
            "type": "object",
            "properties": {
                "file_id": {
                    "type": "string"
                },
                "file_name": {
                    "type": "string"
                },
                "upload_url": {
                    "type": "string"
                }
            }
        },
...
        "internal.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "services.ShareableType": {
            "type": "string",
            "enum": [
                "text",
                "url",
                "file"
            ],
            "x-enum-varnames": [
                "ShareableTypeText",
                "ShareableTypeURL",
                "ShareableTypeFile"
            ]
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "This API uses cookie-based authentication. Clients must send the `access_token` cookie for auth.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        },
        "RefreshTokenAuth": {
            "description": "Refresh token cookie (HttpOnly). Sent only to refresh endpoint.",
            "type": "apiKey",
            "name": "refresh_token",
            "in": "cookie"
        }
    },
...
}